\name{BayesianScreening-package}
\alias{BayesianScreening-package}
\alias{BayesianScreening}
\docType{package}
\title{
Perform Bayesian Screening with Gene Dependence
}
\description{
Includes functions to compute posteriors from a gene-level hierarchical prior, as described in the manuscript "Bayesian genome- and epigenome-wide association studies with gene level dependence."  Also includes functions to perform two-class testing with shared kernels for methylation array data.  The file BinaryExample.r gives example code for simulated categorical data. The file GliomaMethylExample.r gives example code for an application to methylation screening data. 
}
\details{
\tabular{ll}{
Package: \tab BayesianScreening\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2016-01-10\cr
License: \tab GPL-3\cr
}

}
\author{
Eric F. Lock

Maintainer: Eric F. Lock <elock@umn.edu>
}
\references{
Lock, E. F. & Dunson, D. B. (2016). Bayesian genome- and epigenome-wide association studies with gene level dependence. Preprint.
	
Lock, E. F. & Dunson, D. B. (2015). Shared kernel Bayesian screening. Biometrika, 102 (4): 829-842.
}

\keyword{ package }

